Extract invite
